requirements.txt
alembic==1.5.2
appdirs==1.4.4
astroid==2.4.2
blueprint==3.4.2
cffi==1.14.4
click==7.1.2
colorama==0.4.4
config==0.5.0.post0
cryptography==3.3.1
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
flask-blueprint==1.3.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
gunicorn
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
virtualenv==20.4.0
Werkzeug==1.0.1
wrapt==1.12.1
